requirements.txt
cycler==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
navec==0.10.0
numpy==1.21.3
opencv-python==4.5.4.58
pdf2image==1.16.0
Pillow==8.4.0
pyparsing==2.4.7
PyPDF2==1.26.0
pytesseract==0.3.8
python-dateutil==2.8.2
razdel==0.5.0
six==1.16.0
slovnet==0.5.0
